Refactoring for objects/events repository
